# Optimize diffing redo