--- a/samples/Xamarin.Forms/CounterApp/CounterApp/CounterApp.fsproj
+++ b/samples/Xamarin.Forms/CounterApp/CounterApp/CounterApp.fsproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
+        <TargetFramework>netstandard2.1</TargetFramework>
         <EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
     </PropertyGroup>
     <ItemGroup>