--- a/samples/Xamarin.Forms/FabulousContacts/FabulousContacts/Controls/BorderedEntry.fs
+++ b/samples/Xamarin.Forms/FabulousContacts/FabulousContacts/Controls/BorderedEntry.fs
@@ -13,23 +13,26 @@ type BorderedEntry() =
         BindableProperty.Create("BorderColor", typeof<Color>, typeof<BorderedEntry>, Color.Default)
 
     member this.BorderColor
-        with get () =
-            this.GetValue(BorderedEntry.BorderColorProperty) :?> Color
-        and set (value: Color) =
-            this.SetValue(BorderedEntry.BorderColorProperty, value)
+        with get () = this.GetValue(BorderedEntry.BorderColorProperty) :?> Color
+        and set (value: Color) = this.SetValue(BorderedEntry.BorderColorProperty, value)
 
 [<AutoOpen>]
 module FabulousBorderedEntry =
-    let BorderColor = Attributes.defineBindable<Color> BorderedEntry.BorderColorProperty
+    let BorderColor =
+        Attributes.defineBindable<Color> BorderedEntry.BorderColorProperty
+
     let BorderedEntryKey = Widgets.register<BorderedEntry>()
-    
-    type IBorderedEntry = inherit IEntry
+
+    type IBorderedEntry =
+        inherit IEntry
 
     type Fabulous.XamarinForms.View with
         static member inline BorderedEntry<'msg>(text, onTextChanged: string -> 'msg) =
-            ViewHelpers.buildScalars<'msg, IBorderedEntry> BorderedEntryKey
-                [| Entry.Text.WithValue(text)
-                   Entry.TextChanged.WithValue(fun args -> onTextChanged args.NewTextValue |> box) |]
+            WidgetBuilder<'msg, IBorderedEntry>(
+                BorderedEntryKey,
+                Entry.Text.WithValue(text),
+                Entry.TextChanged.WithValue(fun args -> onTextChanged args.NewTextValue |> box)
+            )
 
     [<Extension>]
     type BorderedEntryExtensions =