--- a/samples/Xamarin.Forms/FabulousContacts/FabulousContacts/FabulousContacts.fsproj
+++ b/samples/Xamarin.Forms/FabulousContacts/FabulousContacts/FabulousContacts.fsproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
+        <TargetFramework>netstandard2.1</TargetFramework>
         <EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
     </PropertyGroup>
     <ItemGroup>