--- a/samples/Xamarin.Forms/FabulousWeather/FabulousWeather/FabulousWeather.fsproj
+++ b/samples/Xamarin.Forms/FabulousWeather/FabulousWeather/FabulousWeather.fsproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
+        <TargetFramework>netstandard2.1</TargetFramework>
         <EnableDefaultEmbeddedResourceItems>false</EnableDefaultEmbeddedResourceItems>
     </PropertyGroup>
     <ItemGroup>