--- a/samples/Xamarin.Forms/FabulousWeather/FabulousWeather/PancakeView.fs
+++ b/samples/Xamarin.Forms/FabulousWeather/FabulousWeather/PancakeView.fs
@@ -2,18 +2,35 @@
 
 open Fabulous
 open Fabulous.XamarinForms
+open Fabulous.StackAllocatedCollections
 
 [<AutoOpen>]
 module PancakeView =
-    let BackgroundGradientStops = Attributes.defineBindable<Xamarin.Forms.PancakeView.GradientStopCollection> Xamarin.Forms.PancakeView.PancakeView.BackgroundGradientStopsProperty
-    let Content = Attributes.defineBindableWidget Xamarin.Forms.PancakeView.PancakeView.ContentProperty
-    let PancakeViewKey = Widgets.register<Xamarin.Forms.PancakeView.PancakeView>()
-    
-    type IPancakeView = inherit IView
+    let BackgroundGradientStops =
+        Attributes.defineBindable<Xamarin.Forms.PancakeView.GradientStopCollection>
+            Xamarin.Forms.PancakeView.PancakeView.BackgroundGradientStopsProperty
+
+    let Content =
+        Attributes.defineBindableWidget Xamarin.Forms.PancakeView.PancakeView.ContentProperty
+
+    let PancakeViewKey =
+        Widgets.register<Xamarin.Forms.PancakeView.PancakeView>()
+
+    type IPancakeView =
+        inherit IView
 
     type Fabulous.XamarinForms.View with
-        static member inline PancakeView<'msg, 'marker when 'marker :> IView>(backgroundGradientStops, content: WidgetBuilder<'msg, 'marker>) =
-            ViewHelpers.build<'msg, IPancakeView> PancakeViewKey
-                [| BackgroundGradientStops.WithValue(backgroundGradientStops) |]
-                [| Content.WithValue(content.Compile()) |]
-                [||]
+        static member inline PancakeView<'msg, 'marker when 'marker :> IView>
+            (
+                backgroundGradientStops,
+                content: WidgetBuilder<'msg, 'marker>
+            ) =
+            WidgetBuilder<'msg, IPancakeView>(
+                PancakeViewKey,
+
+                AttributesBundle(
+                    StackArray3.one(BackgroundGradientStops.WithValue(backgroundGradientStops)),
+                    Some [| Content.WithValue(content.Compile()) |],
+                    None
+                )
+            )