--- a/src/Fabulous.Tests/TestUI.Widgets.fs
+++ b/src/Fabulous.Tests/TestUI.Widgets.fs
@@ -3,6 +3,7 @@ module Tests.TestUI_Widgets
 open System
 open System.Runtime.CompilerServices
 open Fabulous
+open Fabulous.StackAllocatedCollections
 open Tests
 open Tests.Platform
 open TestUI_Attributes
@@ -37,10 +38,12 @@ module Widgets =
 
                         view.PropertyBag.Add(ViewNode.ViewNodeProperty, viewNode)
 
+                        let oldWidget: Widget voption = ValueNone
+
                         Reconciler.update
                             //                            context.ViewTreeContext.GetViewNode
                             context.CanReuseView
-                            ValueNone
+                            oldWidget
                             widget
                             viewNode
 
@@ -92,10 +95,16 @@ type WidgetExtensions() =
         this.AddScalar(Attributes.TextStyle.TextColor.WithValue(value))
 
 
-let inline private scalars
-    (value: ScalarAttribute [])
-    : struct (ScalarAttribute [] * WidgetAttribute [] * WidgetCollectionAttribute []) =
-    struct (value, [||], [||])
+    [<Extension>]
+    static member inline record<'msg, 'marker when 'marker :> TestLabelMarker>
+        (
+            this: WidgetBuilder<'msg, 'marker>,
+            value: bool
+        ) =
+        this.AddScalar(Attributes.Text.Record.WithValue(value))
+
+
+let inline private scalars (value: StackArray3<ScalarAttribute>) : AttributesInFlight = struct (value, None, None)
 
 ///----------------
 
@@ -106,18 +115,28 @@ type View private () =
     static let TestStackKey = Widgets.register<TestStack>()
 
     static member Label<'msg>(text: string) =
-        WidgetBuilder<'msg, TestLabelMarker>(TestLabelKey, scalars [| Attributes.Text.Text.WithValue(text) |])
+        WidgetBuilder<'msg, TestLabelMarker>(
+            TestLabelKey,
+            scalars(StackArray3.one(Attributes.Text.Text.WithValue(text)))
+        )
 
 
     static member Button<'msg>(text: string, onClicked: 'msg) =
         WidgetBuilder<'msg, TestButtonMarker>(
             TestButtonKey,
-            scalars [| Attributes.Text.Text.WithValue(text)
-                       Attributes.Button.Pressed.WithValue(onClicked) |]
+            scalars(
+                StackArray3.two(Attributes.Text.Text.WithValue(text), Attributes.Button.Pressed.WithValue(onClicked))
+            )
+        //            scalars [| Attributes.Text.Text.WithValue(text)
+//                       Attributes.Button.Pressed.WithValue(onClicked) |]
         )
 
     static member Stack<'msg, 'marker when 'marker :> IMarker>() =
-        CollectionBuilder<'msg, TestStackMarker, 'marker>(TestStackKey, ValueNone, Attributes.Container.Children)
+        CollectionBuilder<'msg, TestStackMarker, 'marker>(
+            TestStackKey,
+            StackArray3.empty(),
+            Attributes.Container.Children
+        )
 
 [<Extension>]
 type CollectionBuilderExtensions =
@@ -127,7 +146,9 @@ type CollectionBuilderExtensions =
             _: CollectionBuilder<'msg, 'marker, IMarker>,
             x: WidgetBuilder<'msg, 'itemMarker>
         ) : Content<'msg> =
-        { Widgets = [ x.Compile() ] }
+        {
+            Widgets = StackArray3.one(x.Compile())
+        }
 
     [<Extension>]
     static member inline Yield<'msg, 'marker, 'itemMarker when 'itemMarker :> IMarker>
@@ -135,7 +156,9 @@ type CollectionBuilderExtensions =
             _: CollectionBuilder<'msg, 'marker, IMarker>,
             x: WidgetBuilder<'msg, Memo.Memoized<'itemMarker>>
         ) : Content<'msg> =
-        { Widgets = [ x.Compile() ] }
+        {
+            Widgets = StackArray3.one(x.Compile())
+        }
 
     [<Extension>]
     static member inline YieldFrom<'msg, 'marker, 'itemMarker when 'itemMarker :> IMarker>
@@ -144,7 +167,11 @@ type CollectionBuilderExtensions =
             x: WidgetBuilder<'msg, 'itemMarker> seq
         ) : Content<'msg> =
         {
-            Widgets = x |> Seq.map(fun wb -> wb.Compile()) |> Seq.toList
+            Widgets =
+                x
+                |> Seq.map(fun wb -> wb.Compile())
+                |> Seq.toArray
+                |> StackArray3.many
         }
 
 
@@ -199,7 +226,8 @@ module Run =
                 state <- Some(newModel, target, newWidget)
 
                 // ViewNode.getViewNode
-                Reconciler.update x.viewContext.CanReuseView (ValueSome oldWidget) newWidget viewNode
+                let oldWidget = (ValueSome oldWidget)
+                Reconciler.update x.viewContext.CanReuseView oldWidget newWidget viewNode
                 ()
 
         member x.Start(arg: 'arg) =