… src/Fabulous.XamarinForms/WidgetExtensions.fs
… src/Fabulous.XamarinForms/Xamarin.Forms.Core.fs
… src/Fabulous/ArrayExtra.fs
… src/Fabulous/ArraySlice.fs
… src/Fabulous/AttributeDefinitions.fs
… src/Fabulous/Attributes.fs
… src/Fabulous/Builders.fs
… src/Fabulous/Fabulous.fsproj
… src/Fabulous/Memo.fs
… src/Fabulous/Reconciler.fs
… src/Fabulous/StackAllocatedCollections.fs
… src/Fabulous/View.fs
… src/Fabulous/ViewNamespace.fs
… src/Fabulous/ViewNode.fs