@twop 2022-01-09
Used ValueOption instead of option, and used Span instead of ArraySlice when appropriate

here is the latest results (note slight increase in memory consumption but slightly better perf)

|          Method | depth |       Mean |    Error |   StdDev |       Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|---------------- |------ |-----------:|---------:|---------:|------------:|------------:|-----------:|----------:|
| ProcessMessages |    10 |   147.5 ms |  1.66 ms |  1.55 ms |  42750.0000 |  16500.0000 |  1500.0000 |    138 MB |
| ProcessMessages |    15 | 2,910.4 ms | 36.78 ms | 34.40 ms | 476000.0000 | 158000.0000 | 61000.0000 |  1,536 MB |