@TimLariviere 2022-01-09
>here is the latest results (note slight increase in memory consumption but slightly better perf)

Not sure if it's really worth it. We have slightly worse perf and memory on low depth, and barely noticeable win on perf compared to what we will pay in GC pause for bigger depth.

What do you think?