@twop 2022-01-09
> > here is the latest results (note slight increase in memory consumption but slightly better perf)
> 
> Not sure if it's really worth it. We have slightly worse perf and memory on low depth, and barely noticeable win on perf compared to what we will pay in GC pause for bigger depth.
> 
> What do you think?


I think it is, even though it is more memory it is easier GC and faster runtime perf. Note that on M1 cache misses are less severe because of large cache and really wide CPU lines.

So I think it is better, I'm curious how it is going to be on mobile devices